refactor(engine): Phase 4 slice 3 — classifier.py splits into stages/classify/{attributes,subtypes,re_split_flags,permissions,approach,audit}
The 2,227-LOC classifier becomes the mandatory stage anatomy:
attributes.py (cp attribute join + independent flags + the scratch-coupled
expr builders, documented), subtypes.py (SME/retail/QRRE mutation +
exposure subclass), re_split_flags.py (the 6-function RE-split candidate
block in its own module so the Slice-4 re_split decision never moves it
again), permissions.py (model-permission resolution + CLS006 diagnostics),
approach.py (B31 restrictions + the 10-branch approach ladder), audit.py
(audit trail + DQ008/CLS008 collectors), with classify() preserved
VERBATIM on ExposureClassifier (materialise → diagnostics → raw seal
ordering, brand-probe edge selection, has_model_permissions flag flow) and
stage.py the fold adapter. engine/classifier.py survives as a 24-line shim
(logger name pinned); the stale ENTITY_TYPE_TO_* re-export comment is
deleted — verified no importers.

Verbatim move: ratchet metrics byte-identical; zero delegators needed
(no test calls classifier privates). Citation snapshot regenerated
(8 functions re-keyed); Art. 114 citation-hygiene path pin re-pointed at
approach.py; CRR_COVERAGE_NOTES 147A path updated.

Gates: full suite 7,532 passed; parity byte-identical across all four 10k
stress configs; ruff/ty/arch_check green.
